# public/api-spec.yaml
openapi: 3.0.0
info:
  title: Report Mapper API
  version: 1.0.0
  description: API for managing wildlife observation reports with geospatial data
  contact:
    name: Report Mapper Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-production-domain.com/api
    description: Production server
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ${KEYCLOAK_ISSUER}/protocol/openid-connect/auth
          tokenUrl: ${KEYCLOAK_ISSUER}/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email
      description: OAuth2 authentication via Keycloak
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 Bearer token from Keycloak (for direct API access)
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session cookie (automatically set by browser)
  schemas:
    Observation:
      type: object
      properties:
        id:
          type: string
          description: Unique observation identifier
        revisionId:
          type: integer
          description: Revision number
        description:
          type: string
          description: Observation description
        location:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
        imageIds:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              image_key:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        published:
          type: boolean
        submitted:
          type: boolean
        owner:
          type: string
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        canPublish:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
security:
  - oauth2: [openid, profile, email]
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /observations/{id}/revisions:
    post:
      summary: Create a new revision of an existing observation
      description: |
        Create a new revision from an existing observation.

        **Access Control**: Only the owner with edit permission can create a revision.

        **Immutable Properties** (copied from base observation, cannot be changed):
        - observation_id (stays the same across all revisions)
        - created_at (original creation date)
        - owner (ownership cannot be transferred via revisions)

        **Mutable Properties** (can be changed in the new revision):
        - description
        - location
        - imageIds
        - published/submitted status
      tags:
        - Observations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The observation ID to create a revision for
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                imageIds:
                  type: array
                  items:
                    type: string
                publish:
                  type: boolean
                  description: Auto-publish the new revision if user has permission
                submit:
                  type: boolean
                  description: Submit the new revision for review
      responses:
        '201':
          description: New revision created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  revisionId:
                    type: integer
                  published:
                    type: boolean
                  submitted:
                    type: boolean
                  message:
                    type: string
                  immutableProperties:
                    type: object
                    properties:
                      observationId:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      owner:
                        type: string
                  mutableProperties:
                    type: object
                    properties:
                      description:
                        type: string
                      location:
                        type: object
                      imageIds:
                        type: array
        '401':
          description: Authentication required
        '403':
          description: Forbidden - only owner with edit permission can create revisions
        '404':
          description: Base observation not found
  /observations/{id}/{revisionId}:
    get:
      summary: Get a specific revision of an observation
      description: Returns a specific revision by ID
      tags:
        - Observations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Observation revision details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
        '404':
          description: Observation revision not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a specific revision
      description: Update the metadata of a specific revision
      tags:
        - Observations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                imageIds:
                  type: array
                  items:
                    type: string
                publish:
                  type: boolean
                  description: Optionally publish after update
                submit:
                  type: boolean
                  description: Optionally submit for review after update
      responses:
        '200':
          description: Revision updated
        '401':
          description: Authentication required
        '403':
          description: Forbidden
        '404':
          description: Revision not found

    delete:
      summary: Delete a specific revision
      description: Delete a specific revision (requires appropriate permissions)
      tags:
        - Observations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Revision deleted
        '401':
          description: Authentication required
        '403':
          description: Forbidden
        '404':
          description: Revision not found

    patch:
      summary: Publish or submit a specific revision
      description: Change the status of a specific revision
      tags:
        - Observations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [publish, submit]
                  description: Action to perform on the revision
      responses:
        '200':
          description: Action completed successfully
        '401':
          description: Authentication required
        '403':
          description: Forbidden
        '404':
          description: Revision not found
...